move conversion funcs to xrp-utils.ts
